for Jrarat :)

+++ libft/ft_substr.c
@@ -3,24 +3,24 @@
 /*                                                        :::      ::::::::   */
 /*   ft_substr.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: marihovh <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: marihovh <marihovh@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/20 02:20:23 by marihovh          #+#    #+#             */
-/*   Updated: 2023/02/03 22:03:07 by marihovh         ###   ########.fr       */
+/*   Updated: 2023/07/24 14:47:23 by marihovh         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-char	*ft_substr(char const *s, unsigned int start, size_t len)
+char	*ft_substr(char const *s, unsigned int start, int len)
 {
 	char			*sub;
-	size_t			i;
-	size_t			slen;
+	int			i;
+	int			slen;
 
 	i = 0;
 	slen = ft_strlen(s);
-	if (start > slen)
+	if (start > (unsigned)slen)
 		return (ft_strdup(""));
 	if (len > slen)
 		len = slen;
